Made some changes after 1 hour of testing cache features.

+++ atum-jvcp/src/main/java/org/atum/jvcp/cache/CacheExchangeInterface.java
@@ -11,6 +11,6 @@ import org.atum.jvcp.model.EcmRequest;
  */
 public interface CacheExchangeInterface {
 
-	public EcmRequest peekCache(long ecmHash);
-	public boolean addEntry(long ecmHash,EcmRequest dcw);
+	public EcmRequest peekCache(int ecmHash);
+	public boolean addEntry(int ecmHash,EcmRequest dcw);
 }
